Finished calibration, modified motor control to be more adaptable.

+++ 3373Robohawks2016/src/org/usfirst/frc/team3373/robot/HawkCalibration.java
@@ -5,8 +5,8 @@ import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 
 public class HawkCalibration {
 	
-	static double rangeMin;
-	static double rangeMax;
+	static int rangeMin;
+	static int rangeMax;
 	
 	static double calibrationLeftY;
